[REFACTOR] useClipboardRecord 커스텀 훅 로직 생성 및 코드 분리
